bridge/requirements.txt
discord.py==2.4.0
python-telegram-bot==21.6
Telethon==1.36.0
APScheduler==3.10.4
SQLAlchemy==2.0.36
httpx==0.27.2
requests==2.32.3
python-dotenv==1.0.1
tenacity==9.0.0
pydantic==2.9.1
openai==1.54.4
